# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HyperLocal India Super App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: shopId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── Cart ────────────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{cartItemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order from current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details and tracking
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── User Profile ─────────────────────────────────────────────────────────
...
  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit a review or feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackBody"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
  /feedback/product/{productId}:
    get:
      operationId: getProductFeedback
      tags: [feedback]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"

  # ─── Seller Dashboard ─────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Product:
      type: object
      required: [id, name, price, shopId, categoryId, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        mrp:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        shopId:
          type: integer
        shopName:
          type: ["string", "null"]
        categoryId:
          type: integer
        unit:
          type: ["string", "null"]
        inStock:
          type: boolean
        stockQuantity:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        isFeatured:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
    CartItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        shopId:
          type: integer
        shopName:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    Cart:
      type: object
      required: [items, totalItems, totalAmount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalItems:
          type: integer
        totalAmount:
          type: number
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    Order:
      type: object
      required: [id, userId, shopId, status, totalAmount, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        shopId:
          type: integer
        shopName:
          type: ["string", "null"]
        status:
          type: string
          enum: [placed, confirmed, preparing, dispatched, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cod, online]
        paymentStatus:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        deliveryNotes:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Feedback:
      type: object
      required: [id, userId, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        userName:
          type: ["string", "null"]
        productId:
          type: ["integer", "null"]
        shopId:
          type: ["integer", "null"]
        orderId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        type:
          type: string
          enum: [product_review, shop_review, general, complaint]
        createdAt:
          type: string
          format: date-time
...
    AddToCartBody:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    UpdateCartItemBody:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
    PlaceOrderBody:
      type: object
      required: [paymentMethod, deliveryAddress]
      properties:
        paymentMethod:
          type: string
          enum: [cod, online]
        deliveryAddress:
          type: string
        deliveryNotes:
          type: ["string", "null"]
...
    SubmitFeedbackBody:
      type: object
      required: [rating, type]
      properties:
        productId:
          type: ["integer", "null"]
        shopId:
          type: ["integer", "null"]
        orderId:
          type: ["integer", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        type:
          type: string
          enum: [product_review, shop_review, general, complaint]